GraphQLProvider refactoring
The class annotation is changed from @Component to @Configuration since it's more suitable for Bean declaration. It also removes the attribute graphQL: GraphQL since it's no longer necessary.

The sdl is loaded as an InputStream since it's no longer needed to be passed as a String to SchemaParser::parse method.